Add files via upload
At first I just intended to make bouncing balls across my screen but as I began working I wanted to get creative with the visual - and so I tried to make the balls flicker in various colors so I randomised the fill in the ellipses. The problems I encountered were of wanting to incorporate multiple or countless circles automatically instead of manually coding them - so I restricted myself to 5 circles bouncing in random directions, but am keen on learning how I could achieve that.